+++ app/src/main/java/com/example/chotuvemobileapp/SignUpActivity.kt
@@ -13,6 +13,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.example.chotuvemobileapp.data.users.LoginDataSource
 import com.example.chotuvemobileapp.data.users.User
+import com.example.chotuvemobileapp.helpers.Utilities.createDatePicker
 import kotlinx.android.synthetic.main.activity_sign_up.*
 import java.text.DecimalFormat
 import java.time.LocalDate
@@ -24,23 +25,7 @@ class SignUpActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_sign_up)
 
-        val calendar = Calendar.getInstance()
-        val year = calendar.get(Calendar.YEAR)
-        val month = calendar.get(Calendar.MONTH)
-        val day = calendar.get(Calendar.DAY_OF_MONTH)
-
-        RegDateText.setOnClickListener {
-            val dialog = DatePickerDialog(this, DatePickerDialog.OnDateSetListener{ _, mYear, mMonth, mDay ->
-                val df = DecimalFormat("00")
-                val trueMonth = df.format(mMonth + 1)
-                val trueDay = df.format(mDay)
-                val text = "$mYear-$trueMonth-$trueDay"
-                RegDateText.setText(text)
-
-            }, year, month, day)
-            dialog.datePicker.maxDate = Calendar.getInstance().timeInMillis
-            dialog.show()
-        }
+        createDatePicker(RegDateText, this)
 
         SignupProgressBar.visibility = View.GONE
