+++ app/src/main/java/com/example/chotuvemobileapp/ui/profile/ProfileFragment.kt
@@ -34,23 +34,41 @@ class ProfileFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        startFragment()
+    }
+
+    private inner class ScreenSlidePagerAdapter(fa: Fragment) : FragmentStateAdapter(fa) {
+        override fun getItemCount(): Int = 3
+        override fun createFragment(position: Int): Fragment {
+            return when(position){
+                0 -> ProfileDetailsFragment.newInstance(firstName, lastName, email, birthDate, userName)
+                1 -> VideoListFragment.newInstance(3)
+                else -> CommentsFragment.newInstance(19)
+            }
+        }
+    }
+
+    private fun startFragment() {
         ProfileAppbar.alpha = .2F
         ProfileScrollView.alpha = .2F
-        requireActivity().window.setFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE, WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE)
+        requireActivity().window.setFlags(
+            WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE,
+            WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE
+        )
         ProfileProgressBar.visibility = View.VISIBLE
 
         val token = requireActivity().applicationContext
-            .getSharedPreferences(getString(R.string.shared_preferences_file), Context.MODE_PRIVATE).getString("token", "Fail")
+            .getSharedPreferences(getString(R.string.shared_preferences_file), Context.MODE_PRIVATE)
+            .getString("token", "Fail")
         mPager = ProfileViewPager
-        ProfileInfoDataSource.getProfileInfo(token!!){
-            if(it != null){
+        ProfileInfoDataSource.getProfileInfo(token!!) {
+            if (it != null) {
                 firstName = it.first_name
                 lastName = it.last_name
                 userName = it.user_name
                 email = it.email
                 birthDate = it.birthdate
-            }
-            else{
+            } else {
                 firstName = ""
                 lastName = ""
                 userName = ""
@@ -62,7 +80,7 @@ class ProfileFragment : Fragment() {
             UsernameTextView.text = userName
             val pagerAdapter = ScreenSlidePagerAdapter(this)
             mPager.adapter = pagerAdapter
-            TabLayoutMediator(ProfileTabLayout, mPager){ tab, position ->
+            TabLayoutMediator(ProfileTabLayout, mPager) { tab, position ->
                 when (position) {
                     0 -> tab.text = getString(R.string.profile_details)
                     1 -> tab.text = getString(R.string.videos)
@@ -76,14 +94,8 @@ class ProfileFragment : Fragment() {
         }
     }
 
-    private inner class ScreenSlidePagerAdapter(fa: Fragment) : FragmentStateAdapter(fa) {
-        override fun getItemCount(): Int = 3
-        override fun createFragment(position: Int): Fragment {
-            return when(position){
-                0 -> ProfileDetailsFragment.newInstance(firstName, lastName, email, birthDate, userName)
-                1 -> VideoListFragment.newInstance(3)
-                else -> CommentsFragment.newInstance(19)
-            }
-        }
+    override fun onResume() {
+        super.onResume()
+        startFragment()
     }
 }
